Mount the ConfigMap with app metadata in kurl-proxy
This patch also includes improved handling if the app slug was provided or not.
If not, then we'll create an empty ConfigMap (meaning no branding was provided).
If slug provided, then we'll query replicated.app for metadata.  Again, if no
metadata provided, then we'll leave the ConfigMap empty.  If provided, then
we'll populate the ConfigMap with the app's metadata (including branding).

This patch fixes branding on the TLS certs pages.