only prompt to load images when missing
On cluster airgap installs check the list of images on each node's
status to ensure that all kubernetes, docker, and addon images are
available on that node. If a node is found to be missing an image, only
prompt the user to run the load-images task for that node, then
continue on to checking remaining remote nodes.

Addon manifest lines that declare images now have a 4th column
"no_remote_load". There will be no checks to ensure these images exist
on remote nodes. Currently there is a single image using this column,
the bcrypt image in the kotsadm addon which is never used in a pod.