Set X-Kurl-Hash header in script response
The hash is of the installer spec after versions are resolved, which
will allow clients to determine when changes have been made to the spec
of a named installer.
This will allow automated update tools to determine when the script
needs to be re-run to receive upgrades.
This is different than Etag since there are often changes to the scripts
that don't need to be run on every installation.