Suppress key verification warning on airgap RPMs
Prior to this commit, airgap installs would print a key verification
warning for each host RPM installed.  Although the ideal solution would
be to have these RPMs signed such that these warnings do not print, it
would be better to suppress them for now.

With this commit, the warnings are no longer printed as the signature
key is no longer checked.