kubernetes upgrade fails when enabling EKCO with internal load balancer.
When we try to upgarde both kubernetes and EKCO versions in the
same spec along with enabling internal load balancer feature the
upgrade fails as the install.sh script will try check the api
server endpoint of the internal load balacner(localhost:6444) which
has not been depkloyed yet. So, featching the api server endpoint
from the cluster and performing status checks on it during the
upgrade.