Fix unbound variable error on registry backup+restore
This commit updates the registry `backup.sh` and `restore.sh`
scripts to fix a "OBJECT_STORE_HOSTNAME: unbound variable" error
that occurs when restoring registry content from a Velero backup
produced by kURL installations with an first installation date
prior to v2021.12.17-0. Regardless of upgrades, these installations
still carry configuration for `$OBJECT_STORE_CLUSTER_IP` instead of
`$OBJECT_STORE_HOSTNAME` and `set -u` causes the script
to abort when an unset `$OBJECT_STORE_HOSTNAME` is passed to
`test -z`.

The fix is to use a parameter expansion to substitute a zero-length
string if the variable is undefined as per the [Rewriting scripts to
deal with it][1] section of Greg's Bash Guide.

[1]: http://mywiki.wooledge.org/BashFAQ/112