chore: stop saving config + awaiting for pods to finish
after this commit we stopped saving the config when creating the new
minio pvc and the minio backup pvc. if we use save config we can't apply
later because we are hard setting the volume name.

this commit also solves a problem when the minio takes some time to get
back up during an upgrade. we only start the migration once minio is
ready and we can reach it.