chore: read minio store format file inside a loop
when upgrading minio running with longhorn as a backend it can take some
time for the deployment to rollout properly, during this time the pods
are up but we see an io error when attempting to read the file inside of
them. this commit makes the file read process in a loop.