chore: adding test for longhorn migration
added a test to testgrid that validates we can migrate away from
longhorn and into openebs. the test, after the install, creates a
deployment and uploads a file into it (its pods). after the upgrade
the test validate that we can still read the file and that the pvc
was migrated to openebs provisioner.