bug: don't attempt to migrate from longhorn on unsupported version
due to compatibility issues, if the version of Kubernetes being used is
v1.25 or higher, it will not be possible for Longhorn to operate properly
and the migration process will fail. this commit introduces a message that
will be displayed and the process will immediately terminate upon detection
of a Kubernetes version of v1.25 or higher.