chore: standardize minio deployment await timeout.
in some cases testgrid is failing to bring minio up, the failure
reported is:

```
2022-12-28 07:26:30+00:00 awaiting minio deployment
2022-12-28 07:29:00+00:00 [/] An error occurred on line 199
```

by inspecting the testgrid log the pod seems to be running but not in a
healthy state yet:

```
NAMESPACE     NAME                    STATUS    RESTARTS   AGE
minio         minio-5b8df4f997-vhhws  Running   0          2m31s
```

worthy to mention that in the test in case the host preflight had a
warning regarding disk performance (that may be related). as this
problem does not happen consistently (the same test passed in the next
daily run) this commit increases the timeout applied when waiting for
minio to come up.