chore: do not wait for namespace to dissapear.
we have a case testgrid case where the longhorn-system namespace is
taking ages to disappear. this causes the testgrid to timeout after
45 minutes. i have tried this very same test case and i could not
reproduce it.

i have tried it with:

- ubuntu 22.04
- oracle linux 2

both attempts were made using machines with 4 cpus and 16gb of ram,
a configuration that has the same amount of memory and less cpu power
than the testgrid has, in all cases the namespace takes a few seconds
to dissapear.

if i take these two specific attempts at reproducing and add the amount
of times i have manually tested the migration on top without seeing this
issue it is somewhat safe to assume that this might be related to some
hardware constraints in testgrid (maybe when many tests are running at
the same time).

this pr adds the `--wait=false` flag to the longhorn-system delete
operation, allowing the installer to move on and the cluster to delete
the resource on its own time.