feat: allow containerd two minor version upgrades
containerd does not supports upgrades that span more than one minor
release. this pr introduces this control to the installation script.
from now one we only allow containerd upgrades that span two minor
releases and we do one upgrade at a time.

the table below presents some of the outputs of this change:

| from | to   | outcome                                               |
|------|------|-------------------------------------------------------|
| v1.2 | v1.5 | the upgrade is not possible,  error message is shown. |
| v1.3 | v1.4 | upgrade succeeds.                                     |
| v1.3 | v1.5 | upgrades to v1.4 and then to v1.5.                    |
| v1.4 | v1.3 | downgrades are blocked, error message is shown.       |
| v1.6 | v2.0 | major upgrades versions are blocked                   |

about the major version upgrades

as we don't fully understand how the migration to v2 of containerd will
work we have blocked this kind of upgrade. the idea is to make testgrids
to fail in such migrations (once containerd v2 comes to live). we don't
know what kind of upgrade is going to be supported when v2 is released.