feat: check flannel udp port prior to migrating
before migrating from weave to flannel we need to verify if nodes can
communicate with each other through port 8472/udp.

this commit moves this logic to the flannel_pre_init() function, the
sooner the user knows about a problem the better. if the port is not
available the upgrade is aborted.