feat: check flannel udp port prior to migrating (#4380)
* feat: check flannel udp port prior to migrating

before migrating from weave to flannel we need to verify if nodes can
communicate with each other through port 8472/udp.

this commit moves this logic to the flannel_pre_init() function, the
sooner the user knows about a problem the better. if the port is not
available the upgrade is aborted.

* chore: moving nodes connectivity check to a function

* chore: moving containerd_node_is_using_docker() to common.sh

this function is useful in other places as well so it makes sense to
move it to common.sh and rename it (remove the "containerd" prefix).

* chore: loading kurl-util image on local node

we need to load this image otherwise the test will fail on air gapped
enfironments.

* chore: using KURL_UTIL_IMAGE variable.

* chore: stop using deprecated patchesStrategicMerge

patchesStrategicMerge has been deprecated and replaced by "patches".

* bug: wait for job to dissappear before moving on

if we don't wait we may ended up trying to change job's immutable fields
thus trigerring an error.

* chore: using wait.PollUntilContextCancel

wait.PollUntil has been deprecated.