feat: accepts ekco address and number of nodes as flags
since the join.sh script does not have access to a kubeconfig file,
querying the cluster for information such as the number of nodes,
the ekco authentication token, and the ekco pod IP address is not
possible.

with the implementation of this pull request, these parameters will
need to be provided as command line flags to the script.

join.sh will receive them as arguments as well, these are going to
rendered by the install.sh.