chore: inform users that 15% of the root disk is reserved
when evaluating the amount of free disk space to migrate to openebs we
reserve 15% of the disk for the operating system (avoid kubernetes
triggering of its clean up unused images routine). we haven't been
informing users of that causing confusion on why the available + used
amounts did not add up to the full root disk size.