bug: abort immediately if we can't read one of the pod logs
if we are unable to attach to one of the listener pods we abort with an
error. if we don't do this then we were showing up as if the traffic on
the given port was blocked.

this pr solves the following scenario:

1. we deploy the listeners daemonset
2. we wait for the daemonset to rollout.
3. we attempt to attach to listener pod logs (all pods).
4. due to a cluster misconfiguration we can't reach kubelet in one of
   the nodes.

before this pr the step 4 was ignored and we moved on as if we were
attached to the pod, generating a misleading error message.