bug: validate if received data is an uuid
when validating nodes can talk to each other we send uuid from one point
to another but we are not validating if the output matches the expected
uuid. in some scenarios the pod may log some error and if we don't and
on these cases the log is not an uuid and we should retur an error
instead.