bug: stop installing sub-dependencies on rhel-9
on rhel 9 we want to ask user to manually install the dependencies. this
pr also disables the `amazon-id` and the `subscription-manager` yum
plugins when running `yum install`, both plugins attempt to reach the
internet with the former showing up a "stack trace" when it can't reach
the remote servers:

```
Traceback (most recent call last):
  File "/usr/lib/python3.9/site-packages/dnf/plugin.py", line 107, in _caller
    getattr(plugin, method)()
  File "/usr/lib/python3.9/site-packages/dnf-plugins/amazon-id.py", line 111, in config
    aws_region = json.loads(id_doc)['region']
  File "/usr/lib64/python3.9/json/__init__.py", line 339, in loads
    raise TypeError(f'the JSON object must be str, bytes or bytearray, '
TypeError: the JSON object must be str, bytes or bytearray, not NoneType
```

if any of these yum plugins is not installed no side effect has been
noticed during the tests, ie `yum install --disableplugin bogus` does
not generate errors.