bug: retry creating installer for 1m
we have seen cases in our tests where the crd is created but as soon as
we try to create an object it fails. as this happens sometimes and not
all times and judging that we are already using retry for 1m a couple of
lines below let's add the retry here as well.